refactor: seperate PatientScreen to a component
