+++ package.json
@@ -53,7 +53,6 @@
     "@hookform/resolvers": "^3.9.0",
     "@medplum/core": "^4.0.3",
     "@medplum/expo-polyfills": "^4.0.3",
-    "@medplum/fhirtypes": "^4.0.3",
     "@medplum/react-hooks": "^4.0.3",
     "@shopify/flash-list": "1.7.3",
     "@tanstack/react-query": "^5.52.1",
